pyproject.toml
[project]
name = "vinaslt-favorites-service"
version = "0.1.0"
description = ""
authors = [
    {name = "ivanpetrovskij",email = "peyrovskaaa@gmai.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "alembic (>=1.17.0,<2.0.0)",
    "fastapi (>=0.119.0,<0.120.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "grpcio (>=1.75.1,<2.0.0)",
    "protobuf (>=6.33.0,<7.0.0)",
    "fastapi-pagination (>=0.14.3,<0.15.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "authtools @ git+https://github.com/vinasLT/shared_auth.git@v0.1.5",
    "factory-boy (>=3.3.3,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
